backend/requirements.txt
# Core Backend Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Cloud Storage
cloudinary>=1.36.0

# Database - PostgreSQL with Vector Support
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pgvector>=0.2.4
alembic>=1.13.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML Models
transformers>=4.35.0
# NOTE: For reproducible CPU-only installs use the PyTorch index URL:
#   pip install torch==2.9.1+cpu torchvision==0.24.1+cpu torchaudio==2.9.1+cpu --index-url https://download.pytorch.org/whl/cpu
torch==2.9.1+cpu
torchvision==0.24.1+cpu
torchaudio==2.9.1+cpu
open-clip-torch>=2.24.0

# Image Processing
Pillow>=10.1.0
opencv-python-headless>=4.8.1
scikit-image>=0.22.0
matplotlib>=3.8.2

# PDF Processing
pdfplumber>=0.10.3
PyMuPDF>=1.23.8

# Scientific Computing
numpy>=1.26.2
scipy>=1.11.4

# Utilities
python-dotenv>=1.0.0

# Additional for deployment
gunicorn>=21.2.0
requests>=2.31.0

# Streamlit UI
streamlit

# MySQL Support (for legacy WhatsApp integration)
mysql-connector-python

# WhatsApp Integration
twilio
flask

# AI Chat Support
google-generativeai
openai
